LayoutTests/http/wpt/beacon/connect-src-beacon-redirect-blocked.sub-expected.txt


PASS Redirect is blocked by CSP 
